add build

+++ src/cars-service/dto/CarDTO.cpp
@@ -72,9 +72,9 @@ std::string ToJSON(const CarsDTO &cars, size_t page, size_t pageSize)
 {
     Json::Value root;
 
-    root["page"] = page;
-    root["pageSize"] = pageSize;
-    root["totalElements"] = cars.size();
+    root["page"] = (uint) page;
+    root["pageSize"] = (uint) pageSize;
+    root["totalElements"] = (uint) cars.size();
 
     Json::Value items = Json::arrayValue;
     for (auto car : cars)
